backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
mysql-connector-python==9.1.0
python-dotenv==1.0.1
bcrypt==4.2.0
PyJWT==2.9.0
flask-jwt-extended==4.6.0
